Fix DbEnv initialization and add clang warning compatibility check

+++ src/db.cpp
@@ -43,10 +43,10 @@ void CDBEnv::EnvShutdown()
     {
         printf("EnvShutdown exception: %s (%d)\n", e.what(), e.get_errno());
     }
-    DbEnv(0).remove(GetDataDir().string().c_str(), 0);
+    DbEnv(u_int32_t(0)).remove(GetDataDir().string().c_str(), 0);
 }
 
-CDBEnv::CDBEnv() : dbenv(0)
+CDBEnv::CDBEnv() : dbenv(u_int32_t(0))
 {
 }
